shooter/requirements.txt
zerorpc
gevent
python-logstash